Reorganize remaining default miner modules (#131)
